refactor(openehr-adl): extract the shared substrate (aom, odin, hrid, artefact) and unify duplicated helpers
Give the crate-wide substrate a single home each: `aom/{access,build,interval}`
(the 13-arm C_OBJECT field accessors + AomType, the AOM2 constructors, and the
Bounds/multiplicity + INTERVAL<T> maths), `artefact` (ArchetypeView/view +
ArchetypeRepository/FlatParent, which also kills the flatten/opt/meta -> validate
inverted dependency), `hrid` (parse/print/lookup-key of ARCHETYPE_HRID), `odin`
(the ODIN reading bridge, master03 escape decoding, delimited-regex handling),
and `paths::child_path`. Duplicate families merged: child_occurrences x4,
sibling_order x3, strip_sibling_order x2, common_mut x2, bounds x2, the
finite-upper trio, the ODIN key->String trio, untyped x2, nil_uuid x2,
child_path x5, cobject_to_primitive x2, and the AOM primitive classification
(three inlined matches now call aom_type().is_primitive()).

Two divergent pairs are co-located rather than merged, each call site left on
its original variant with a TODO naming the tracking issue: the interval
point-of extractors (point_interval_value_i32 vs degenerate_point_value_i32,
#1339) and the 4-digit-only \uHHHH escape decoding (#1340). The two bounds
renderers and the two delimited-regex predicates keep distinct names plus doc
comments stating the difference, because both spellings feed user-visible
message text. Behaviour-preserving: no test expectation or corpus edits.
