+++ crates/openehr-adl/src/lib.rs
@@ -11,7 +11,12 @@
 //! Components: the [`lexer`], the outer artefact parser ([`source`]), the
 //! shared typed [`error`] catalogue, and the cADL definition-section parser
 //! ([`cadl`]) that builds the generated `openehr_am::am24::aom2` constraint
-//! model; [`codes`]/[`paths`] (code math + ADL paths), [`validate`] (the AOM2
+//! model. Under all of it sits the shared substrate every component reads
+//! through: [`aom`] (AOM2 field accessors, constructors, interval arithmetic),
+//! [`artefact`] (artefact views + the parent/supplier repository), [`hrid`]
+//! (the archetype-id grammar), and [`odin`] (the ODIN reading bridge + the
+//! `master03` lexical decoding). Above it:
+//! [`codes`]/[`paths`] (code math + ADL paths), [`validate`] (the AOM2
 //! validation catalogue), [`flatten`] + [`opt`] (specialisation flattening +
 //! OPT2), the [`printer`], and [`adl14`] (ADL 1.4→2 conversion — our own
 //! design, no openEHR spec). ODIN sections are parsed by the
@@ -24,13 +29,17 @@
 // (C-QUESTION-MARK, https://rust-lang.github.io/api-guidelines/documentation.html#c-question-mark).
 #![doc(test(attr(deny(warnings))))]
 pub mod adl14;
+pub mod aom;
+pub mod artefact;
 pub mod assemble;
 pub mod cadl;
 pub mod codes;
 pub mod error;
 pub mod flatten;
+pub mod hrid;
 pub mod lexer;
 pub mod meta;
+pub mod odin;
 pub mod opt;
 pub mod paths;
 pub mod printer;
