+++ crates/openehr-adl/tests/it/corpus_validity_phase1.rs
@@ -18,8 +18,9 @@
 use std::collections::BTreeMap;
 use std::path::{Path, PathBuf};
 
+use openehr_adl::artefact::ArchetypeRepository;
 use openehr_adl::assemble::parse_artefact;
-use openehr_adl::validate::{ArchetypeRepository, Severity, validate_source_phase1};
+use openehr_adl::validate::{Severity, validate_source_phase1};
 
 const CORPUS: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/tests/corpus/adl2-reference");
