+++ crates/openehr-adl/tests/it/validation_phase1_cases.rs
@@ -18,9 +18,10 @@
     reason = "test assertions panic/unwrap/expect by design, including in fixture helpers outside #[test] fns"
 )]
 
+use openehr_adl::artefact::ArchetypeRepository;
 use openehr_adl::assemble::parse_artefact;
 use openehr_adl::validate::{
-    ArchetypeRepository, ValidationCode, ValidationIssue, validate_phase1, validate_source_phase1,
+    ValidationCode, ValidationIssue, validate_phase1, validate_source_phase1,
 };
 use openehr_am::am24::aom2::archetype::archetype::Archetype;
 use openehr_am::am24::aom2::archetype::authored_archetype::{
